[sockjs] Delete sockjs, rename sockjs-node to sockjs, add tslint.json
@types/sockjs-node was impossible to use since the actual name of the
package it's meant for on npm is sockjs, not sockjs-node. @types/sockjs
already existed, however it was the type declaration file for an old
version of sockjs-client, which already has its own up-to-date type
declaration file under @types/sockjs-client.

Fixes #17112
